Fixed 41 "-Wpointer-to-int-cast" warnings
by introducing the usage of uintptr_t to many places where a pointer needed to be treated as a numerical value, and vice-a-versa.

+++ src/r_drawflush.inl
@@ -28,6 +28,8 @@
  *
  *-----------------------------------------------------------------------------*/
 
+#include <inttypes.h>
+
 #if (R_DRAWCOLUMN_PIPELINE_BITS == 8)
 #define SCREENTYPE byte
 #define TOPLEFT byte_topleft
@@ -239,8 +241,8 @@ static void R_FLUSHQUAD_FUNCNAME(void) {
   }
 #else
 #if (R_DRAWCOLUMN_PIPELINE_BITS == 8)
-  if ((sizeof(int) == 4) && (((int)source % 4) == 0) &&
-      (((int)dest % 4) == 0)) {
+  if ((sizeof(int) == 4) && (((uintptr_t)source % 4) == 0) &&
+      (((uintptr_t)dest % 4) == 0)) {
     while (--count >= 0) {
       *(int *)dest = *(int *)source;
       source += 4 * sizeof(byte);
